+++ test/coveredVaultFactory.ts
